Refactor Schedulable and DependantItems implementations in modal dialogs #469
