+++ src/main/resources/assets/js/app/issue/view/IssueDetailsDialog.ts
@@ -1,4 +1,3 @@
-import {SchedulableDialog} from '../../dialog/SchedulableDialog';
 import {Issue} from '../Issue';
 import {ContentPublishPromptEvent} from '../../browse/ContentPublishPromptEvent';
 import {Router} from '../../Router';
@@ -19,6 +18,7 @@ import {IssueCommentTextArea} from './IssueCommentTextArea';
 import {CreateIssueCommentRequest} from '../resource/CreateIssueCommentRequest';
 import {IssueDetailsDialogHeader} from './IssueDetailsDialogHeader';
 import {PublishContentRequest} from '../../resource/PublishContentRequest';
+import {BasePublishDialog} from '../../dialog/BasePublishDialog';
 import AEl = api.dom.AEl;
 import DialogButton = api.ui.dialog.DialogButton;
 import ContentSummaryAndCompareStatusFetcher = api.content.resource.ContentSummaryAndCompareStatusFetcher;
@@ -44,7 +44,7 @@ import ComboBox = api.ui.selector.combobox.ComboBox;
 import ContentId = api.content.ContentId;
 
 export class IssueDetailsDialog
-    extends SchedulableDialog {
+    extends BasePublishDialog {
 
     private issue: Issue;
