+++ src/main/resources/assets/js/app/publish/SchedulePublishDialog.ts
@@ -1,42 +1,30 @@
 import AEl = api.dom.AEl;
 import i18n = api.util.i18n;
+import FormItem = api.form.FormItem;
 import {PublishFrom} from '../inputtype/publish/PublishFrom';
 import {PublishToFuture} from '../inputtype/publish/PublishToFuture';
+import {SchedulableDialog} from '../dialog/SchedulableDialog';
 
 export class SchedulePublishDialog
-    extends api.ui.dialog.ModalDialog {
-
-    private formView: api.form.FormView;
-    private propertySet: api.data.PropertySet;
-
-    private confirmDeleteButton: api.ui.dialog.DialogButton;
-
-    private confirmScheduleAction: api.ui.Action;
+    extends SchedulableDialog {
 
     private onCloseCallback: () => void;
-
-    private onScheduleCallback: () => void;
 
     constructor() {
         super(<api.ui.dialog.ModalDialogConfig>{
             title: i18n('dialog.schedule')
         });
-
         this.getEl().addClass('schedule-publish-dialog');
 
         this.addSubtitle();
-
-        this.initConfirmScheduleAction();
-
-        this.initFormView();
 
         this.createBackButton();
     }
 
     show() {
-        this.resetPublishDates();
+        this.resetDates();
         this.formView.displayValidationErrors(false);
-        this.confirmScheduleAction.setEnabled(true);
+        this.confirmScheduleButton.getAction().setEnabled(true);
         super.show();
     }
 
@@ -52,35 +40,22 @@ export class SchedulePublishDialog
         this.onCloseCallback = onCloseCallback;
     }
 
-    onSchedule(onScheduleCallback: () => void) {
-        this.onScheduleCallback = onScheduleCallback;
+    protected createFromFormItem(): FormItem {
+        return new api.form.InputBuilder().setName('from').setInputType(PublishFrom.getName()).setLabel(
+            i18n('field.onlineFrom')).setHelpText(i18n('field.onlineFrom.help')).setOccurrences(
+            new api.form.OccurrencesBuilder().setMinimum(1).setMaximum(1).build()).setInputTypeConfig({}).setMaximizeUIInputWidth(
+            true).build();
     }
 
-    private initFormView() {
-        let formBuilder = new api.form.FormBuilder().addFormItem(
-            new api.form.InputBuilder().setName('from').setInputType(PublishFrom.getName()).setLabel(
-                i18n('field.onlineFrom')).setHelpText(i18n('field.onlineFrom.help')).setOccurrences(
-                new api.form.OccurrencesBuilder().setMinimum(1).setMaximum(1).build()).setInputTypeConfig({}).setMaximizeUIInputWidth(
-                true).build()).addFormItem(
-            new api.form.InputBuilder().setName('to').setInputType(PublishToFuture.getName()).setLabel(
-                i18n('field.onlineTo')).setHelpText(i18n('field.onlineTo.help')).setOccurrences(
-                new api.form.OccurrencesBuilder().setMinimum(0).setMaximum(1).build()).setInputTypeConfig({}).setMaximizeUIInputWidth(
-                true).build());
+    protected createToFormItem(): FormItem {
+        return new api.form.InputBuilder().setName('to').setInputType(PublishToFuture.getName()).setLabel(
+            i18n('field.onlineTo')).setHelpText(i18n('field.onlineTo.help')).setOccurrences(
+            new api.form.OccurrencesBuilder().setMinimum(0).setMaximum(1).build()).setInputTypeConfig({}).setMaximizeUIInputWidth(
+            true).build();
+    }
 
-        this.propertySet = new api.data.PropertyTree().getRoot();
-        this.formView = new api.form.FormView(api.form.FormContext.create().build(), formBuilder.build(), this.propertySet);
-
-        this.formView.layout().then(() => {
-            this.formView.onValidityChanged((event: api.form.FormValidityChangedEvent) => {
-                this.confirmScheduleAction.setEnabled(event.isValid());
-                this.formView.displayValidationErrors(true);
-            });
-            this.propertySet.onChanged(() => {
-                this.formView.validate();
-            });
-            this.appendChildToContentPanel(this.formView);
-
-        });
+    protected hasSubDialog(): boolean {
+        return true;
     }
 
     private createBackButton() {
@@ -96,45 +71,5 @@ export class SchedulePublishDialog
     private addSubtitle() {
         this.appendChildToHeader(
             new api.dom.H6El('schedule-publish-dialog-subtitle').setHtml(i18n('dialog.schedule.subname'), false));
-    }
-
-    private initConfirmScheduleAction() {
-        this.confirmScheduleAction = new api.ui.Action(i18n('action.schedule'));
-
-        this.confirmScheduleAction.setIconClass('confirm-schedule-action');
-        this.confirmScheduleAction.onExecuted(() => {
-            let validationRecording = this.formView.validate();
-            if (validationRecording.isValid()) {
-                this.close();
-                if (this.onScheduleCallback) {
-                    this.onScheduleCallback();
-                }
-            } else {
-                this.confirmScheduleAction.setEnabled(false);
-                this.formView.displayValidationErrors(true);
-            }
-        });
-
-        this.confirmDeleteButton = this.addAction(this.confirmScheduleAction, true, true);
-    }
-
-    getFromDate(): Date {
-        let publishFrom = this.propertySet.getDateTime('from');
-        return publishFrom && publishFrom.toDate();
-    }
-
-    getToDate(): Date {
-        let publishTo = this.propertySet.getDateTime('to');
-        return publishTo && publishTo.toDate();
-    }
-
-    resetPublishDates() {
-        this.propertySet.reset();
-        this.formView.update(this.propertySet);
-        this.formView.validate();
-    }
-
-    protected hasSubDialog(): boolean {
-        return true;
     }
 }
